lignée d'araignée au pas ! enlever quelques inlines -> plus grande rapidité de compilation (on rajoutera après si jamais) enlever qq qDebugs
git-svn-id: http://www.genecraft-project.com/svn/genecraft/trunk@88 2ce2bb41-2ff2-4fd5-96a4-e3c75a7bd90e

+++ src/genelab/core/entities/brain/synapse.cpp
@@ -15,7 +15,7 @@ namespace GeneLabCore {
         foreach(QVariant v, l) {
             int x = v.toMap()["x"].toInt();
             int y = v.toMap()["y"].toInt();
-            float w = v.toMap()["w"].toFloat();
+            float w = v.toMap()["w"].toDouble();
             this->connectTo(x, y, w);
         }
